Fixed some features and errors

+++ app/src/main/java/com/example/intrack/DistanceCalculator.java
@@ -1,12 +1,7 @@
 package com.example.intrack;
 
 public class DistanceCalculator {
-
-    // This method takes 3 values:
-    // 1. The live RSSI from the scan
-    // 2. The reference RSSI at 1 meter
-    // 3. The environment factor (n)
-    public static double calculateDistance(int rssi, int rssiAtOneMeter, double pathLossExponent) {
-        return Math.pow(10.0, ((rssiAtOneMeter - rssi) / (10.0 * pathLossExponent)));
+    public static double calculateDistance(int rssi, double rssiAtOneMeter, double pathLossExponent) {
+        return Math.pow(10.0, (rssiAtOneMeter - rssi) / (10 * pathLossExponent));
     }
 }
